requirements.txt
aiogram>=3.0.0
flask>=2.3.0
requests>=2.31.0
pymongo>=4.5.0
pillow>=10.0.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
# и другие, которые вы используете
